Split ps.telescope module into init, mappings, setup
`mappings` is written in such a way to allow other modules (specifically
to be able to set telescope mapping only when a lsp has been attached
using the `on_attach` functionallity of lspconfg). to set buffer
specific mappings for telescope. Though, we would first have to define
the corresponding function in `init`