kitty: stop updating tab bar every second
There's no point since the bar stops updating as soon as I launch nvim.
In any case, the tab update as soon as it detects any activity be it
moving of your mouse or pressing any key.

I'll just to have to remember to press a key in order to see the latest
changes.